+++ internal/provider/provider.go
@@ -243,6 +243,7 @@ func (p *SdwanProvider) Resources(ctx context.Context) []func() resource.Resourc
 	return []func() resource.Resource{
 		NewCEdgeAAAFeatureTemplateResource,
 		NewCEdgeGlobalFeatureTemplateResource,
+		NewCellularProfileFeatureTemplateResource,
 		NewCiscoBannerFeatureTemplateResource,
 		NewCiscoBFDFeatureTemplateResource,
 		NewCiscoBGPFeatureTemplateResource,
@@ -341,6 +342,7 @@ func (p *SdwanProvider) DataSources(ctx context.Context) []func() datasource.Dat
 	return []func() datasource.DataSource{
 		NewCEdgeAAAFeatureTemplateDataSource,
 		NewCEdgeGlobalFeatureTemplateDataSource,
+		NewCellularProfileFeatureTemplateDataSource,
 		NewCiscoBannerFeatureTemplateDataSource,
 		NewCiscoBFDFeatureTemplateDataSource,
 		NewCiscoBGPFeatureTemplateDataSource,